requirements.txt
# OverView V0.3 依赖

# === 数据库 ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# === Supabase ===
supabase>=2.0.0

# === 爬虫 ===
selenium>=4.0.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# === LLM API ===
openai>=1.0.0
json-repair>=0.1.0

# === PDF 处理 ===
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# === Docling (PDF OCR 支持) ===
docling>=2.0.0
# 可选 OCR 引擎（根据需要安装）:
# - easyocr (默认，已包含在 docling 中)
# - tesseract: 需要系统安装 tesseract-ocr
# - rapidocr: pip install rapidocr_onnxruntime

# === DOC 处理 ===
python-docx>=1.0.0

# === 工具 ===
requests>=2.31.0
python-dotenv>=1.0.0
